@@ -1,16 +1,16 @@
 """Custom exception classes for the application."""
 
-from typing import Any, Dict, Optional
+from typing import Any
 
 
 class APIException(Exception):
     """Base exception class for API-related errors."""
-    
+
     def __init__(
-        self, 
+        self,
         message: str,
         status_code: int = 500,
-        details: Optional[Dict[str, Any]] = None
+        details: dict[str, Any] | None = None
     ):
         self.message = message
         self.status_code = status_code
@@ -20,41 +20,41 @@ class APIException(Exception):
 
 class ValidationError(APIException):
     """Exception raised for validation errors."""
-    
-    def __init__(self, message: str, details: Optional[Dict[str, Any]] = None):
+
+    def __init__(self, message: str, details: dict[str, Any] | None = None):
         super().__init__(message, status_code=400, details=details)
 
 
 class NotFoundError(APIException):
     """Exception raised when a resource is not found."""
-    
-    def __init__(self, message: str = "Resource not found", details: Optional[Dict[str, Any]] = None):
+
+    def __init__(self, message: str = "Resource not found", details: dict[str, Any] | None = None):
         super().__init__(message, status_code=404, details=details)
 
 
 class AuthenticationError(APIException):
     """Exception raised for authentication errors."""
-    
-    def __init__(self, message: str = "Authentication failed", details: Optional[Dict[str, Any]] = None):
+
+    def __init__(self, message: str = "Authentication failed", details: dict[str, Any] | None = None):
         super().__init__(message, status_code=401, details=details)
 
 
 class AuthorizationError(APIException):
     """Exception raised for authorization errors."""
-    
-    def __init__(self, message: str = "Access denied", details: Optional[Dict[str, Any]] = None):
+
+    def __init__(self, message: str = "Access denied", details: dict[str, Any] | None = None):
         super().__init__(message, status_code=403, details=details)
 
 
 class ConflictError(APIException):
     """Exception raised for resource conflicts."""
-    
-    def __init__(self, message: str = "Resource conflict", details: Optional[Dict[str, Any]] = None):
+
+    def __init__(self, message: str = "Resource conflict", details: dict[str, Any] | None = None):
         super().__init__(message, status_code=409, details=details)
 
 
 class DatabaseError(APIException):
     """Exception raised for database-related errors."""
-    
-    def __init__(self, message: str = "Database operation failed", details: Optional[Dict[str, Any]] = None):
+
+    def __init__(self, message: str = "Database operation failed", details: dict[str, Any] | None = None):
         super().__init__(message, status_code=500, details=details)
