@@ -1,13 +1,12 @@
-"""Test configuration and fixtures with PostgreSQL integration."""
+"""Test configuration and fixtures with SQLite integration."""
 
 import pytest
 import asyncio
-import uuid
-from typing import AsyncGenerator, Generator
+import tempfile
+import os
+from typing import AsyncGenerator
 from fastapi.testclient import TestClient
-from sqlalchemy import create_engine, text
 from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine, async_sessionmaker
-from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
 from app.models.base import Base
@@ -18,30 +17,15 @@ from app.models.user import User
 from app.core.security import get_password_hash
 from main import app
 
-# PostgreSQL test database configuration
-POSTGRES_USER = "postgres"
-POSTGRES_PASSWORD = "mydbpassword"
-POSTGRES_HOST = "127.0.0.1"
-POSTGRES_PORT = "5432"
-POSTGRES_DB_BASE = "fastapi_test"
+# Test database configuration using SQLite
+TEST_DATABASE_URL = "sqlite+aiosqlite:///./test.db"
 
-# Use a simpler approach with per-function database isolation
-POSTGRES_URL = f"postgresql://{POSTGRES_USER}:{POSTGRES_PASSWORD}@{POSTGRES_HOST}:{POSTGRES_PORT}"
-POSTGRES_TEST_DB = f"{POSTGRES_DB_BASE}_main"
-TEST_DATABASE_URL = f"{POSTGRES_URL}/{POSTGRES_TEST_DB}"
-ASYNC_TEST_DATABASE_URL = f"postgresql+asyncpg://{POSTGRES_USER}:{POSTGRES_PASSWORD}@{POSTGRES_HOST}:{POSTGRES_PORT}/{POSTGRES_TEST_DB}"
-
-# Create sync engine for database setup/teardown
-setup_engine = create_engine(f"{POSTGRES_URL}/postgres", isolation_level="AUTOCOMMIT")
-
-# Async engine for actual testing with smaller pool to avoid connection issues
+# Create test engine with proper SQLite configuration
 async_test_engine = create_async_engine(
-    ASYNC_TEST_DATABASE_URL,
+    TEST_DATABASE_URL,
     echo=False,
-    pool_size=5,
-    max_overflow=10,
-    pool_pre_ping=True,
-    pool_recycle=300
+    connect_args={"check_same_thread": False},
+    poolclass=StaticPool,
 )
 
 AsyncTestingSessionLocal = async_sessionmaker(
@@ -56,57 +40,40 @@ AsyncTestingSessionLocal = async_sessionmaker(
 @pytest.fixture(scope="session")
 def event_loop():
     """Create an instance of the default event loop for the test session."""
-    policy = asyncio.get_event_loop_policy()
-    loop = policy.new_event_loop()
+    loop = asyncio.new_event_loop()
     yield loop
     loop.close()
 
 
 @pytest.fixture(scope="session", autouse=True)
-def setup_test_database():
-    """Create test database for the session."""
-    # Create the test database
-    with setup_engine.connect() as conn:
-        conn.execute(text(f"DROP DATABASE IF EXISTS {POSTGRES_TEST_DB}"))
-        conn.execute(text(f"CREATE DATABASE {POSTGRES_TEST_DB}"))
-    
-    yield
-    
-    # Cleanup: Drop test database after all tests
-    with setup_engine.connect() as conn:
-        # Force disconnect any remaining connections
-        conn.execute(text(f"""
-            SELECT pg_terminate_backend(pid) 
-            FROM pg_stat_activity 
-            WHERE datname = '{POSTGRES_TEST_DB}' AND pid <> pg_backend_pid()
-        """))
-        conn.execute(text(f"DROP DATABASE IF EXISTS {POSTGRES_TEST_DB}"))
-
-
-@pytest.fixture(scope="session")
-async def setup_tables():
-    """Create tables after database is created."""
+async def setup_test_database():
+    """Create and setup test database tables."""
+    # Create all tables
     async with async_test_engine.begin() as conn:
         await conn.run_sync(Base.metadata.create_all)
     
     yield
     
-    # Dispose of all connections
+    # Cleanup: Drop all tables and dispose engine
+    async with async_test_engine.begin() as conn:
+        await conn.run_sync(Base.metadata.drop_all)
     await async_test_engine.dispose()
-
+    
+    # Remove test database file if it exists
+    if os.path.exists("./test.db"):
+        os.remove("./test.db")
 
 
 @pytest.fixture
-async def async_db_session(setup_tables) -> AsyncGenerator[AsyncSession, None]:
-    """Create async database session for integration tests with transaction rollback."""
+async def async_db_session() -> AsyncGenerator[AsyncSession, None]:
+    """Create async database session for tests with transaction rollback."""
     async with AsyncTestingSessionLocal() as session:
-        # Start a transaction that we can rollback to isolate tests
-        transaction = await session.begin()
         try:
             yield session
         finally:
-            # Always rollback to isolate tests
-            await transaction.rollback()
+            # Rollback any uncommitted changes
+            if session.in_transaction():
+                await session.rollback()
             await session.close()
 
 
@@ -127,8 +94,8 @@ def override_get_user_service(async_db_session: AsyncSession):
 
 
 @pytest.fixture
-async def client_with_real_db(async_db_session: AsyncSession):
-    """Create test client with real database integration."""
+async def client_with_db(async_db_session: AsyncSession):
+    """Create test client with database integration."""
     # Override dependencies
     app.dependency_overrides[get_async_db] = lambda: async_db_session
     app.dependency_overrides[get_user_service] = lambda: UserService(async_db_session)
@@ -142,17 +109,19 @@ async def client_with_real_db(async_db_session: AsyncSession):
 
 @pytest.fixture
 def client():
-    """Create simple test client for basic API tests (legacy compatibility)."""
+    """Create simple test client for basic API tests."""
     with TestClient(app) as test_client:
         yield test_client
 
 
 @pytest.fixture
-async def sample_user_in_db(async_db_session: AsyncSession):
-    """Create a real user in the test database."""
+async def sample_user(async_db_session: AsyncSession):
+    """Create a test user in the database."""
+    import uuid
+    unique_id = str(uuid.uuid4())[:8]
     user_data = User(
-        username="testuser",
-        email="test@example.com", 
+        username=f"testuser_{unique_id}",
+        email=f"test_{unique_id}@example.com", 
         full_name="Test User",
         hashed_password=get_password_hash("TestPass123!"),
         is_active=True,
@@ -167,11 +136,13 @@ async def sample_user_in_db(async_db_session: AsyncSession):
 
 
 @pytest.fixture
-async def admin_user_in_db(async_db_session: AsyncSession):
-    """Create a real admin user in the test database."""
+async def admin_user(async_db_session: AsyncSession):
+    """Create an admin user in the database."""
+    import uuid
+    unique_id = str(uuid.uuid4())[:8]
     admin_data = User(
-        username="admin",
-        email="admin@example.com",
+        username=f"admin_{unique_id}",
+        email=f"admin_{unique_id}@example.com",
         full_name="Admin User", 
         hashed_password=get_password_hash("AdminPass123!"),
         is_active=True,
@@ -186,23 +157,62 @@ async def admin_user_in_db(async_db_session: AsyncSession):
 
 
 @pytest.fixture
-async def auth_headers(client_with_real_db, admin_user_in_db):
-    """Get real authentication headers using OAuth2 login."""
-    login_response = client_with_real_db.post(
+async def auth_headers(client_with_db, admin_user):
+    """Get authentication headers using OAuth2 login."""
+    # Use the dynamic admin user's email
+    admin_email = admin_user.email
+    
+    # First try the local login endpoint
+    login_response = client_with_db.post(
         "/api/v1/oauth/login",
         json={
-            "email": "admin@example.com",
+            "email": admin_email,
             "password": "AdminPass123!",
             "grant_type": "password"
         }
     )
-    assert login_response.status_code == 200
-    token = login_response.json()["access_token"]
-    return {"Authorization": f"Bearer {token}"}
+    
+    if login_response.status_code == 200:
+        token = login_response.json()["access_token"]
+        return {"Authorization": f"Bearer {token}"}
+    
+    # If OAuth login doesn't work, try the basic auth login
+    login_response = client_with_db.post(
+        "/api/v1/auth/login",
+        data={
+            "username": admin_email,
+            "password": "AdminPass123!"
+        }
+    )
+    
+    if login_response.status_code == 200:
+        token = login_response.json()["access_token"]
+        return {"Authorization": f"Bearer {token}"}
+    
+    # If both fail, raise an error with debugging info
+    raise Exception(f"Login failed for {admin_email}. OAuth response: {login_response.status_code} - {login_response.text}")
 
 
-# Legacy fixtures for backward compatibility with existing tests
+# Backward compatibility fixtures
 @pytest.fixture
 async def async_session(async_db_session: AsyncSession):
     """Legacy fixture name for backward compatibility."""
     return async_db_session
+
+
+@pytest.fixture
+async def sample_user_in_db(sample_user: User):
+    """Legacy fixture name for backward compatibility."""
+    return sample_user
+
+
+@pytest.fixture
+async def admin_user_in_db(admin_user: User):
+    """Legacy fixture name for backward compatibility."""
+    return admin_user
+
+
+@pytest.fixture
+async def client_with_real_db(client_with_db):
+    """Legacy fixture name for backward compatibility."""
+    return client_with_db
