@@ -5,18 +5,18 @@ from fastapi.security import OAuth2PasswordBearer
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core.database import get_async_db
-from app.core.security import verify_token
 from app.core.exceptions import AuthenticationError
-from app.services.user import UserService
-from app.services.oauth import GoogleOAuthProvider
+from app.core.security import verify_token
 from app.models.user import User
+from app.services.oauth import GoogleOAuthProvider
+from app.services.user import UserService
 
 # OAuth2 scheme
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/oauth/login")
 
 
 async def get_current_user(
-    token: str = Depends(oauth2_scheme), 
+    token: str = Depends(oauth2_scheme),
     db: AsyncSession = Depends(get_async_db)
 ) -> User:
     """Get current authenticated user supporting both local JWT and Google ID tokens."""
@@ -28,7 +28,7 @@ async def get_current_user(
 
     try:
         user_service = UserService(db)
-        
+
         # First try to verify as local JWT token
         payload = verify_token(token)
         if payload is not None:
@@ -42,12 +42,12 @@ async def get_current_user(
                         return user
             except (ValueError, Exception):
                 pass  # Fall through to try Google token
-        
+
         # Try to verify as Google ID token
         try:
             oauth_service = GoogleOAuthProvider()
             id_info = await oauth_service.validate_id_token(token)
-            
+
             # Extract Google user ID
             google_user_id = id_info.get("sub")
             if google_user_id:
@@ -56,10 +56,10 @@ async def get_current_user(
                     return user
         except Exception:
             pass  # Not a valid Google token
-        
+
         # If we get here, token validation failed
         raise credentials_exception
-        
+
     except AuthenticationError:
         raise credentials_exception
     except Exception:
