@@ -1,8 +1,8 @@
 """Health check endpoints."""
 
 from fastapi import APIRouter, Depends
-from sqlalchemy.orm import Session
 from sqlalchemy import text
+from sqlalchemy.orm import Session
 
 from app.core.database import get_db
 
