@@ -1,15 +1,15 @@
 """Enhanced user management endpoints with pagination and search."""
 
-from typing import Any, List
+from typing import Any
 
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 
-from app.api.dependencies import get_user_service, get_current_user, get_current_active_user
-from app.services.user import UserService
+from app.api.dependencies import get_current_active_user, get_user_service
+from app.core.exceptions import ConflictError, NotFoundError, ValidationError
 from app.models.user import User
+from app.schemas.pagination import PaginatedResponse, PaginationParams
 from app.schemas.user import UserCreate, UserResponse, UserUpdate
-from app.schemas.pagination import PaginationParams, PaginatedResponse
-from app.core.exceptions import ValidationError, NotFoundError, ConflictError
+from app.services.user import UserService
 
 router = APIRouter()
 
