@@ -1,11 +1,11 @@
 """OAuth service providers package."""
 
 from app.services.oauth.base import BaseOAuthProvider
-from app.services.oauth.google import GoogleOAuthProvider
 from app.services.oauth.factory import OAuthProviderFactory
+from app.services.oauth.google import GoogleOAuthProvider
 
 __all__ = [
     "BaseOAuthProvider",
-    "GoogleOAuthProvider", 
+    "GoogleOAuthProvider",
     "OAuthProviderFactory",
 ]
