@@ -6,7 +6,7 @@ from typing import Any
 
 class BaseOAuthProvider(ABC):
     """Abstract base class for OAuth providers."""
-    
+
     @abstractmethod
     async def get_authorization_url(
         self,
@@ -17,7 +17,7 @@ class BaseOAuthProvider(ABC):
     ) -> str:
         """Generate OAuth authorization URL."""
         pass
-    
+
     @abstractmethod
     async def exchange_code_for_tokens(
         self,
@@ -27,17 +27,17 @@ class BaseOAuthProvider(ABC):
     ) -> dict[str, Any]:
         """Exchange authorization code for tokens."""
         pass
-    
+
     @abstractmethod
     async def validate_id_token(self, id_token: str) -> dict[str, Any]:
         """Validate and decode ID token."""
         pass
-    
+
     @abstractmethod
     async def get_user_info(self, access_token: str) -> dict[str, Any]:
         """Get user information from provider."""
         pass
-    
+
     @abstractmethod
     async def refresh_access_token(self, refresh_token: str) -> dict[str, Any]:
         """Refresh access token using refresh token."""
